# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MathPlatform API",
        "description": "API for MathPlatform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/auth/activate": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Submit a student activation request",
                "description": "Creates a student, enrollment, subscription, and subscription code. The code remains hidden until admin approval.",
                "operationId": "activateStudentAccount",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone",
                                    "parent_phone",
                                    "grade_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ahmed"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "01012345678"
                                    },
                                    "parent_phone": {
                                        "type": "string",
                                        "example": "01112345678"
                                    },
                                    "grade_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Activation request created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Your activation request has been submitted successfully. Please wait for admin approval."
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "pending"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Student login",
                "description": "Logs in an approved student using their subscription code and automatically detects the device information from the request.",
                "operationId": "studentLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "subscription_code",
                                    "device_identifier"
                                ],
                                "properties": {
                                    "subscription_code": {
                                        "type": "string",
                                        "example": "A7K9P2X4LM"
                                    },
                                    "device_identifier": {
                                        "description": "Unique device identifier generated by the client application.",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "8f4c2e91-7a32-4d8b-91a2-5e6f3c8d1204"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid subscription code"
                    },
                    "403": {
                        "description": "Subscription is not approved, active, or another device is already trusted"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Student logout",
                "description": "Logs out the authenticated student by deleting the current Sanctum token. The trusted device remains registered.",
                "operationId": "studentLogout",
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the Sanctum token returned by the admin login.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}